[Refactor] Redefine `Direction` and `Action` as enums
This patch refactors the `hypatia/constants.py` file.  The purpose of
the file/module is to define constants Hypatia uses for representing
directions an actor may face or move, and actions they may perform,
such as standing or walking.  However, despite the module's name, the
"constants" are not actually constant.  It is possible for errant code
to create a bug with something like:

    constants.Up = "Broken"

Admittedly this is a far-fetched and unlikely bug; such a line of code
would not be written directly.  But indirect redefinitions of the
constants are a possibility.

This patch eliminates the potential for this bug by redefining
directions and actions in terms of the `Enum` class.  It removes the
sub-classes of `Direction` and `Action` since those classes are now
represented as enumerated members of their parents.  This means we
must change every use of `constants` throughout the code-base, which
this patch does, but it makes it impossible for code like

    constants.Direction.Up = "Broken"

to run without raising an error.  The new constants have arbitrary
integer values but they are not derived from `IntEnum` because we have
no reason to ever compare them to actual integers, i.e. there is no
scenario where we would want to write such conditions as

    if constants.Action.Walk == 1: ...

The `Enum` class was introduced by Python 3.4, so in order for Hypatia
to remain compatible with Python 2.7 it is necessary to install the
`enum34` package, linked below, for systems not using Python 3.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

Resolves: #26

See-Also: https://docs.python.org/3/library/enum.html
See-Also: https://pypi.python.org/pypi/enum34