[Feature] Implement an Actor class
A common design in game engines, particularly for role-playing games,
is to represent many entities as "actors", objects which share
functionality common to anything that can "act" in the game world.
For example, in role-playing games all monsters, NPCs, and the player
character are actors.  They share certain behavior such as being able
to move around the game world, and often share common types of data
such as hit-points, status effects, and so on.

This commit the Actor class to fill this role in the engine.  It is
possible in the future that this should be an abstract base class,
unable for direct instantiation.  However, this patch does not take
that approach because it is too early to tell whether or not the
engine will have useful, necessary reasons for wanting to directly
instantiate objects of the Actor class.

This patch changes the Player class to inherit from Actor, as it is
currently the one class in the engine that will benefit from this type
of change in architecture.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

See-Also: http://rmvxace.wikia.com/wiki/Actors_and_Classes
See-Also: https://docs.unrealengine.com/latest/INT/Programming/UnrealArchitecture/Actors/index.html
See-Also: https://github.com/renpy/renpy/blob/master/renpy/character.py
See-Also: https://github.com/ejmr/LNVL/blob/master/src/character.lua