[Refactor] Move project meta-information into `hypatia/__init__.py`
A number of files have repeated information about the author, contact
information, development status, and so on.  Being in multiple files
means it is likely this information will become inconsistent during
the course of development, e.g. someone could forget to properly
update the `__status__` in a module.  This patch addresses the issue
by moving all of that meta-information into the `__init__.py` file for
the `hypatia` library.

This patch also deletes the `hypatia/version.py` file and puts the
`__version__` information in the same file with all of the similar
information.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>