[Feature] Add a class for representing the current version
This patch implements the `Version` class inside of the `__init__.py`
file alongside the rest of Hypatia's meta-data.  The class has public
integer properties representing the major, minor, and patch portions
of the version number.  This makes it possible for other code in the
engine to support or deprecate features based on the version by
comparing their numeric values, e.g.

    if Version.major < 3:
        # Report some error about a deprecated feature.

The `Version` class also implements `__str__()` so that it is possible
to convert `__version__` into a string, e.g. for `print()`, for the
purpose of output, e.g.

    print(__version__)    # "0.2.8"

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>