[Test] Add a pre-push Git hook to automatically run the test suite
This patch adds a `pre-push` shell script for use as a Git hook.  When
installed it will cause git-push to reject pushing anything which
causes the Hypatia test suite to fail.  The script is inside a new
`etc/' directory because it is not required for either using Hypatia
or developing it; being a mere optional tool for developers it seems
appropriate to place it inside a directory for such optional scripts.

Even after installing the script it is always possible to push without
invoking it by running

    $ git push --no-verify ...

The patch comes with a README in the `etc/` directory which contains
more information about the hook.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

See-Also: http://githooks.com/
See-Also: https://github.com/git/git/blob/master/Documentation/githooks.txt#L181