[Fix] Do not let `Direction` enums share values
This patch redefines the cardinal directions in the `Direction` enum
to use values based on powers of two.  Most notably we change the
value of `Direction.west` from six to eight.  This change is because
of the following condition:

    Direction.west == Direction.south_east

This condition is true when `Direction.west` is equal to six.
Redefining the cardinal directions as powers of two avoids any
potential problems by which members of `Direction` share the same
values.  We define the addition operator for `Direction` which makes
this change useful for another reason, because now the values for all
ordinal directions correctly equal the addition of their cardinal
components, e.g. this is now true:

    Direction.east + Direction.north == Direction.north_east

The definition for `Direction.__add__()` now has a doctest
demonstrating this property.

To safeguard against similar problems in the future we wrap the
definitions of both enum classes, i.e. `Direction` and `Action`, with
the `enum.unique` decorator, which will raise an error if we ever try
to define properties of the enums that have the same values.

Finally this patch corrects a few random typos.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

See-Also: https://docs.python.org/3/library/enum.html?#ensuring-unique-enumeration-values