[Feature] test.sh: Make the version of Python configurable
This patch introduces an optional environment variable into the
`test.sh` shell script: `$HYPATIA_PYTHON_VERSION`.  The purpose of
this variable is to deal with the following situation...

I keep multiple versions of Python installed on my computer: 2.7, 3.4,
and 3.5.  These different versions all have varying names for their
respective executbles, e.g. `python` versus `python3` versus
`python3.5`, and the same for `pip`, where `pip3` represents a
different version.  The `test.sh` script uses the following two
programs in its commands:

    1. `python`
    2. `pip`

There is no guarantee what versions of Python these programs represent
when using these command names.  For example, on my computer `python`
runs Python 2.7, but for another developer it may run Python 3.  There
is no way to tell `test.sh` to use a specific version of Python and
PIP without modifying the script directly.  This is undesirable
because, for example, I may temporarily change the script to use
Python and PIP 3.5 and then accidentally forget to reset the changes,
and thus accidentally commit them to the repository which may screw up
the script for other developers.

The new `$HYPATIA_PYTHON_VERSION` environment variable addresses this
issue by allowing developers to optionally specify which version of
Python and PIP the test script should use.  For example, by setting
the value of `$HYPATIA_PYTHON_VERSION` to "3.5" the script will use
these programs:

    1. `python3.5`
    2. `pip3.5`

If the variable does not exist or is not set then the script defaults
to using `python` and `pip` like it always has, meaning this change is
backwards-compatible.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>