[Feature] Tell Git to ignore PKG-INFO
The `test.sh` script creates the `PKG-INFO` but then deletes it.
However, if the test script stops prematurely then there is a chance
the `PKG-INFO` file will remain.  This patch adds `PKG-INFO` to the
list of files for Git to ignore so that we do not accidentally commit
it in the event that `test.sh` stops before deleting that file.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>