[Feature] Support adding and subtracting tuples with Anchor objects
Anchors are essentially wrappers around 2-tuples of integers,
representing X-Y coordinates.  However, if we want to add an 2-tuple
of coordinates to an anchor we must write the following code:

    >>> our_anchor = Anchor(10, 20)
    >>> our_tuple  = (4, 8)
    >>> result     = our_anchor + Anchor(*our_tuple)
    <Anchor at (14, 28)>

In other words, we must coerce the tuple into an `Anchor` object first
before we can perform the addition.  This patch eliminates that
coercion by rewritting `Anchor.__add__()` and `Anchor.__sub__()` so
that they accept a union of anchors and 2-tuples of integers.  This
allows us to rewrite the above like so:

    >>> our_anchor = Anchor(10, 20)
    >>> our_tuple  = (4, 8)
    >>> result     = our_anchor + our_tuple
    <Anchor at (14, 28)>

We have gained nothing in terms of lines-of-code, but we save
ourselves a coercion and reduce the need to construct a second
`Anchor` object to perform the addition.  And all of this is true for
subtraction as well.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>