[Feature] Allow Anchor objects to be right-hand operands in math
This patch builds on 247e485b ('Support adding and subtracting tuples
with Anchor objects') by implementing two new methods for the `Anchor`
class:

1. `Anchor.__radd__()`
2. `Anchor.__rsub__()`

These methods allow anchors to be the right-hand operands when
performing addition or subtraction between `Anchor` objects and
2-tuples of integers, e.g. we can now write this code:

    >>> coordinates = (100, 200)
    >>> anchor = Anchor(1, 2)
    >>> coordinates + anchor
    <Anchor at (101, 202)>
    >>> coordinates - anchor
    <Anchor at (99, 198)>

The implementation of `__radd__()` is a simple one-liner since
addition for natural numbers is commutative.  However, note well the
code for `__rsub__()`, which must take care to deal with the fact that
subtraction is anticommutative without causing an infinite loop by
swapping the positions of the operands.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>