[Feature] Support integer multiplication for Anchor objects
This patch allows us to write code like this:

    >>> Anchor(3, 5) * -1
    <Anchor at (-3, -5)>

The benefit of this functionality is it provides us with a way to
"scale" or "magnify" an anchor.  It also lets us simplify the
definition of the `__rsub__()` method, which this patch does.

This patch does not implement support for multiplying `Anchor` objects
together.  Currently there is no obvious benefit to allow multiplying
one anchor to another, and it is not obvious what exactly would be the
desirable result of such multiplication.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>