[Feature] Raise NotImplemented errors in Anchor's arithmetic methods
Recent commits have extended the support of arthimetic for the
`Anchor` class.  However, none of the methods follow the advice of the
link below by raising `NotImplemented` exceptions when attempting
arithmetic with unsupported types of values.  For example, we have no
reason to allow the addition of complex numbers and anchors, but if we
write...

    Anchor(1, 3) + complex(2, 4)

...we will, at best, end up with an anchor that has bizarre X-Y
coordinates, and that will very likely break code elsewhere.

This patch addresses the problem by performing type-checking in all of
the arithmetic methods to ensure that they receive the correct types
of arguments.  Those methods will raise the `NotImplemented` error if
we try to write code like the example above.  This helps us discover
mistakes more easily, e.g. if we accidentally multiply an anchor
together with a floating-point number.

At this point it is worth explaining specifically why we do not
support floats.  An anchor represents a position on a surface of
pixels.  We do not allow the use of floating-point numbers because it
does not make sense for an anchor to cover half a pixel, one-third of
a pixel, or anything except an entire pixel.  Therefore we only allow
integers in the arithmetic methods.

Finally, lets discuss performance.  For the sake of example we will
use the following two functions:

    def foo(x, y):
        return x * y

    def bar(x, y):
        if type(x) == int and type(y) == int:
            return x * y
        else:
            raise NotImplemented

Here are the bytecodes for `foo()`:

    0 LOAD_FAST                0 (x)
    3 LOAD_FAST                1 (y)
    6 BINARY_MULTIPLY
    7 RETURN_VALUE

And here are the bytecodes for `bar()`:

     0 LOAD_GLOBAL              0 (type)
     3 LOAD_FAST                0 (x)
     6 CALL_FUNCTION            1 (1 positional, 0 keyword pair)
     9 LOAD_GLOBAL              1 (int)
    12 COMPARE_OP               2 (==)
    15 POP_JUMP_IF_FALSE       44
    18 LOAD_GLOBAL              0 (type)
    21 LOAD_FAST                1 (y)
    24 CALL_FUNCTION            1 (1 positional, 0 keyword pair)
    27 LOAD_GLOBAL              1 (int)
    30 COMPARE_OP               2 (==)
    33 POP_JUMP_IF_FALSE       44

    36 LOAD_FAST                0 (x)
    39 LOAD_FAST                1 (y)
    42 BINARY_MULTIPLY
    43 RETURN_VALUE

    44 LOAD_GLOBAL              2 (NotImplemented)
    47 RAISE_VARARGS            1
    50 LOAD_CONST               0 (None)
    53 RETURN_VALUE

Calling `foo()` generates four bytecodes while `bar()` generates
nineteen.  It is also important to note that `foo()` only needs to use
the `LOAD_FAST` bytecode, while `bar()` has three uses of the bytecode
`LOAD_GLOBAL`.  It is obvious which of the two functions is slower.
However, the most crucial aspect is that we do not know how much
slower `bar()` is compared to the `foo()` function.

This patch, by introducing the use of `NotImplemented`, similarly
increases the number of bytecodes for the arithmetic methods for
`Anchor` objects.  But again, we do not know how much those extra
bytecodes affect performance.  Therefore, due to a lack of profiling
data we will not second-guess ourselves or make assumptions about the
performance implications of this patch.  To put it another way, our
approach is in line with the famous quote by Donald Knuth:

    "Premature optimization is the root of all evil."

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

See-Also: https://docs.python.org/3/library/numbers.html#implementing-the-arithmetic-operations
See-Also: http://c2.com/cgi/wiki?PrematureOptimization