[Test] Skip doctests for the Anchor class which involve exceptions
Commit 37db06cf ("Raise NotImplemented errors in Anchor's arithmetic
methods") introduced doctests for the exceptions raised by the changes
in that patch.  They work for fine Python 3, but break the test suite
when using Python 2.  The problem is that Python 2 raises slightly
different exceptions, usually with different error messages than what
the doctests expect, causing them to fail.

This patch adds the `+SKIP` property to all of the relevant doctests;
see the links below for more information.  The result is that the test
suite will skip those tests for all versions of Python.  This is
undesirable simply because it removes the greatest benefit of having
the doctests in the first place.  I would just out-right delete them.
However, they do still have one minor benefit, which is providing some
documentation on the behavior of the code.  So for that reason this
patch leaves the tests in place but tells Python to always skip them.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #90
See-Also: https://pymotw.com/2/doctest/
See-Also: https://docs.python.org/3/library/doctest.html?highlight=doctest#doctest.SKIP