New contributing guidelines and other corrections to `CONTRIBUTING.md`.
Update `CONTRIBUTING.md` to reflect new contributor standards as talked about in GitHub issue #87, entitled "Process (Documentation, Conceptualization, etc)."

Introduces a talking point for changes to be made to the contributor rules and guidelines. This exists in a separate branch, so we all may discourse on the matter of changes and how they've been implemented.

I took the liberty of adding my own notes and vision as well.

GitHub-Issue: #87