[Feature] Make Actor.NoResponseReason a unique enum
It is unlikely we would ever duplicate one of the strings which are
the values of this enum, but copy-paste errors happen when you're
writing code at 2:00 in the morning and paying less than half
attention to what you're doing.  So making the enum unique safeguards
us from creating any duplicate messages by mistake.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>