[Feature] Introduce a parent class for all Actor-related exceptions
Currently we have one exception related to actors: `NoResponse`.
There are two problems with the exception:

1. The name is ambiguous.  When we raise a `NoResponse` error is it
   because an NPC failed to respond to the player? Or did the server fail
   to respond properly to a networking client. Or did a concurrent
   process never respond to some request?

2. The exception inherits directly from `Exception`.  This creates a
   "shallow" hierarchy for exceptions, which is to say the fact that
   it is a child of `Exception` gives us no additional information
   about its purpose or intent.

This patch address the first issue by renaming the exception from
`NoResponse` to `NoActorResponse`.  It address the second issue by
introducing a new exception: `ActorException`.  This new exception
inherits from `Exception` and becomes the new parent class of our
`NoActorResponse` exception.  Not only does this provide more clarity
about the intend of child exceptions, it paves the way in the future
to introduce additional information and functionality to all
Actor-related exceptions if necessary, which we would not want to bolt
directly onto Python's own `Exception` class.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>