[Fix] Raise AttributeError's when actors have no 'walkabout' property
The implementation of the `Actor.direction` property uses
`walkabout.direction` behind the scenes, but it is possible for an
actor to have a value of None for `walkabout`.  Which breaks access to
the `direction` property.  This patch deals with the issue by raising
`AttributeError` exceptions when we try to do anything with the
property `direction` in situations where there is no walkabout.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>