[Fix] Do not let actors talk() when they have no direction property
The implementation of `Actor.talk()` hinges on the presence of the
`direction` property, which itself comes from the `walkabout`
property.  However, it is acceptable to create actors which have no
value for the `walkabout` property, and such actors break when calling
the `talk()` method.  This patch address the problem by testing for
the presence of `walkabout` in the `talk()` method.  If it does not
exist then the code raises a new exception introduced by this patch:
`ActorCannotTalk`.  We could use `AttributeError` instead but we use a
custom exception because there can be other reasons an actor may be
unable to talk, reasons which may be unrelated to any attributes.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>