[Feature] Use the $PAGER environment variable in test.sh
Most POSIX-based platforms make use of the `$PAGER` environment
variable.  For example, the `man` command uses the program named by
`$PAGER` to display manual pages.

The `test.sh` script always uses `more`.  This patch changes the
script so that it will use the uesr's personal `$PAGER` if it is set.
And if not set the script defaults to using `more` like before.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>