2 Anchor math type-checking fixes
ejmr's previous pull commit (GitHub-Issue: #90) implemented Anchor multiplication, but the reason for these changes is the type handling for Anchor calculations. Anchors can _do math_ with two other types--a tuple of two integers _and_ another Anchor object. The first issue in this implementation was attempting to literally `raise NotImplemented`, the correction was to use the `NoteImplementedError` as so: `raise NotImplementedError(coordinates)`. This correction improves the exception raised by _working_ and supplying what triggered the exception (the coordinates).

In a separate, but related, issue I goof'd by using floats when doing anchor calculations for blitting Walkabout children, namely thanks to `topleft_float` being the primary attribute referenced for `Walkabout` sprite position. So, for example, something like (53.38, 4.03) was being passed for pixel calculations, which is why the exception was being raised in the first place (both NotImplementedError and the bug that started this change). For now I used `int()` to convert the floats from `topleft_float` within `Walkabout.blit()` which is the source of our Anchor woes.

The original message which alerted me to this bug:

> *TypeError:* exceptions must be old-style classes or derived from BaseException, not NotImplementedType

GitHub-Issue: #90