[Feature] Implement an 'active' property for the NPC class
The ultimate goal is to allow for the creation of `NPC` objects which
are usable only once, e.g. a door which the player can "use" only once
by opening it, after which we permanently deactivate the object; see
the discussion for the relevant GitHub issue for a visual example.

This patch introduces the `active` property to the `NPC` class as the
first step towards that goal.  If an NPC is active, i.e. its `active`
property has the boolean True value, then other actors in the game
world may interact with that NPC and vice versa.  If the NPC has a
False `activate` property then it cannot affect the game state and
nothing can interact with it.  This patch does not implement that
logic though.  Instead it simply implements the `active` property
itself along with the necessary documentation.  We do not allow code
to delete the property since we must always be able to determine if an
NPC is active or not, and so trying to delete the property will always
raise a `TypeError` exception.

By default the `active` property is False.  This allows us to create a
large number of NPCs while remaining selective about which we
specifically wish to activate.  This approach is more manageable than
if we created a lot of NPCs that were active by default and had to
de-activate all of those which were not necessary at the time.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #59