[Feature] Implement NPC.on_activation() and NPC.on_deactivation()
These two methods are fundamental to the `active` property of the
`NPC` class and its objects.  Whenever we change the `active` property
on an NPC that object will invoke one of these two methods, e.g.:

    >>> npc = NPC()

    # Invokes npc.on_activation()
    >>> npc.active = True

    # Invokes npc.on_deactivation()
    >>> npc.active = False

These methods allow us to implement custom behavior for NPCs that the
class will automatically trigger whenever we activate or deactivate an
NPC object.  The methods have no-op implementations.  This is because
we cannot predict what is the best default behavior for either method,
so we leave it up to subclasses of `NPC` to implement the appropriate
behavior for the type of NPC they represent.  See the docstring for
the `NPC` class for an example.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #59