[Feature] Support str()-conversion for NPC objects
This patch provides a simple implementation of the `__str__()` method
for `NPC` objects which we can use for debugging purposes.  The string
version of an NPC simply tells us whether or not the NPC is active,
based on the value of its `active` property.  We can use this in logs
of debugging output by calling `print()` on NPCs to see whether or not
they are active or unactive as we expect, which will be useful for
debugging sub-classes that implement specialized `on_activation()` and
`on_deactivation()` methods.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #59