[Test] Add unit tests for the NPC class
This patch adds a new file, `tests/test_player.py`, which contains
unit tests for the `NPC` class, intended for use with pytest.  One of
the unit tests requires the `unittest.mock` package in order to test
whether or not certain methods were called as expected.  That package
is part of the standard library in Python 3.4 and later.  However, in
order to support Python 2.7 we must add the `mock` package to the list
of requirements.  In order to use the package in conjunction with
pytest we must also install the `pytest-mock` module.  So this patch
updates the appropriate files in the `requirements/` directory so that
the test process will install both packages as needed.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #59
See-Also: https://pypi.python.org/pypi/mock
See-Also: https://github.com/pytest-dev/pytest-mock