[Refactor] User super() in the constructors for HumanPlayer and NPC
This patch replaces the explicit calls to `Actor.__init__()` with uses
of `super().__init__()`.  This not only simplifies the code but
protects us from errors in the future if we change the class
hierarchy for actors, NPCs, and so on.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

See-Also: https://docs.python.org/3/library/functions.html#super
See-Also: http://rhettinger.wordpress.com/2011/05/26/super-considered-super/