Platform-specific pre-requisite instructions
Installing Pygame on OS X and Windows can be tricky, so I provided sections on getting Pygame working on either platform. This branch/commit started initially as an assurance that Hypatia works on OS X (it does!).

I added the Windows and OS X Pygame install notes in `README.md`.