[Fix] Correct unit tests for Actor-related exceptions
The unit tests for the `Actor` class include tests which raise
exceptions.  However, these tests raise `NoResponseReason`, which is
not an exception class.  Instead they should raise `NoActorResponse`
and instantiate it with a value from the `NoResponseReason` enum.
This patch modifies the tests so that they raise the correct
exceptions, and with this patch applied all tests in the file
`tests/test_actor.py` pass.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

Fixes: #93