setup.py handles essential dependencies, more docs
Previously, to install the depends for Hypatia, a user had to `pip install -r requirements/base.txt` for Python 3 and `pip install -r requirements/python2.txt` for Python 2. The aforementioned is counter-intuitive and is a truly unneccessary step.

This commit changes `setup.py` so it installs essential dependencies according to user's Python version. The files `requirements/python2.txt` and `requirements/base.txt` have been removed.

The reason for the complicated dependency management is that trying to install `enum34` will crash `setup.py` in versions of Python 3.4 or higher. `enum34` is a backport of the `enum` feature introduced in Python 3.4.

Changes made according to new dependency installation changes (taking out references to the old requirements files). Such as `README.md` instructions and the `script/bootstrap` script. The bootstrap script simply does not attempt to install dependencies.

I also removed some outdated information in the main comment of `test.sh`

`pygame` is still not auto-installed... :<

Add details on distributing and installing to the `setup.py` modue-level docstring

GitHub-Issue: #91