[Style] Correct all currently existing PEP8 style problems
This patch modifies a lot of file but does not change any logic in any
of the code.  All the patch does is reformat certain code so that it
conforms to PEP8.  Since the test suite checks all files for PEP8
compliance this patch eliminates all PEP8-related warnings.

All of the stylistic problems this patch addresses are all similar in
that they relate to the indentation and layout of function calls,
tuples, and dictionaries over multiple lines.  For example, this is a
common pattern:

    foo = (
            bar,
            baz,
            qux,
          )

The test suite sees such code as an error with regard to conforming to
PEP8.  This patch re-arranges all such code like so:

    foo = (bar,
           baz,
           qux)

A similar problem appears with the use of dictionaries.  For example:

    dispositions = {
                    # Cardinal directions' dispositions
                    Direction.north: (0, -margin),
                    Direction.east: (margin, 0),
                    Direction.south: (0, margin),
                    Direction.west: (-margin, 0),
                    # Ordinal directions' dispositions
                    Direction.north_east: (margin, -margin),
                    Direction.south_east: (margin, margin),
                    Direction.south_west: (-margin, margin),
                    Direction.north_west: (-margin, -margin),
                   }

This example cannot be rewritten exactly like the first example due to
the inline comment; PEP8 will complain about it's indentation if the
comment appears immediately after the opening brace.  So this patch
reformats all such code like so:

    dispositions = {
        # Cardinal directions' dispositions
        Direction.north: (0, -margin),
        Direction.east: (margin, 0),
        Direction.south: (0, margin),
        Direction.west: (-margin, 0),

        # Ordinal directions' dispositions
        Direction.north_east: (margin, -margin),
        Direction.south_east: (margin, margin),
        Direction.south_west: (-margin, margin),
        Direction.north_west: (-margin, -margin)
    }

This format satisfies PEP8, but it also has the additional benefit of
allowing us to keep comments aligned with the dictionary entries and
to add extra whitespace above comments as-needed to visually separate
sections of the dictionary entries.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>
Reviewed-by: Lily Lemmer <lily.lemmer@ave81.com>

See-Also: https://www.python.org/dev/peps/pep-0008/