[Feature] Support limitations on activating NPCs
The over-arching goal of issue #59 and this branch is to make it
possible for Hypatia to declare NPCs as "usable once".  This commit
provides that functionality by giving NPC objects a new property:
`activation_limit`.  This limit affects how many times the code can
activate an NPC, i.e. set its `active` property to True.  Therefore we
can now describe an NPC as "usable once" by writing code like so:

    npc = NPC()
    npc.activation_limit = 1

    # Now we can write the following only once...
    npc.active = True

Building on the example above, if at any point we attempt to set
`npc.active` to True a second time the code will raise a new
exception: `CannotActivateNPC`, with an error message mentioning the
activation limit.  Semantically this NPC is "usable once", achieving
the overall goal of this branch.

This patch adds one more relevant public property to the NPC class:
`activation_count`, the number of times we have activated the NPC.
The engine can access this property any time it needs to see how many
times an NPC has been activated.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #59