[Feature] Safeguard NPC.activation_count from external manipulation
The number of times we activate an NPC is information which should
only be modifiable by the actual NPC object we're working with.
Particularly tricky bugs could appear if we allow code like this:

    npc = NPC()
    npc.active = True

    # Now `npc.activation_count == 1` but we can "lie" and break the
    # system by writing this:
    npc.activation_count = 0

This patch redefines `activation_count` as a property which will raise
a TypeError exception in the situation above, or if we ever tried to
out-right delete the property.  This doesn't make it impossible to
manipulate the activation counter from outside of the class, but it
does make it difficult enough that the odds of a creating a bug with
code like the above example are significantly lower.

Finally, the patch adds an additional test to the test suite
demonstrating these new restrictions and slightly increasing the code
coverage for `hypatia/player.py`.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>

GitHub-Issue: #59